[REVIEW] Update share_normal_form for lazy prop
Eager application of contracts acted directly on the annotation attached
to a field. Those were thus transformed by the share_normal_form to make
it work with recursive overriding.

Now, runtime contracts are all conained in the new pending_contracts
field. We can let metadata alone (that is, we don't have to transform
them anymore), while we have to transform pending_contracts, which is
what this commit does.