[REVIEW] Fix pending contracts not dropped + tests
Fix failing tests. The first one was due to pending contracts not being
dropped after being applied inside `%record_map%`, which means that the
function was "sandwiched" by the original contract.

The second one was just due to some bugs in the test itself and in the
pretty printer. Those code paths were just triggered by the
implementation of lazy propagation by accident.